+++ packages/shared-components/src/lib/@types/types.d.ts
@@ -1,3 +1,2 @@
 export * from '@testing-library/jest-dom';
 export * from 'jest-styled-components';
-declare module '*.svg';