+++ packages/shared-components/src/lib/abstracts/colors/index.ts
@@ -1,5 +1,0 @@
-export * from './colors';
-export * from './colorTheme';
-export * from './cssColorVariables';
-export * from './palette';
-export * from './types';