+++ packages/shared-components/src/lib/abstracts/index.ts
@@ -1,7 +1,3 @@
-export * from './breakpoints';
-export { colors } from './colors';
-export * from './elevation';
 export * from './icons';
 export * from './layoutFoundation.styles';
 export { GlobalStyle } from './styles/GlobalStyle';
-export * from './typography';