+++ packages/shared-components/src/lib/abstracts/styles/GlobalStyle.ts
@@ -7,7 +7,7 @@ import '../icons/style.css';
 // TODO: create a generic global style css file in @mordech/tokens
 export const GlobalStyle = createGlobalStyle`
   :root {
-    --bounce-transition: 250ms cubic-bezier(0.65, -1.63, 0.28, 2.72);
+    --mrd-bounce-transition: 250ms cubic-bezier(0.65, -1.63, 0.28, 2.72);
     font-size: ${fontSizes[1]};
     font-weight: ${fontWeights['regular']};
     line-height: 1.25;