+++ packages/shared-components/src/lib/components/ArticleSpotlight/stories/ArticleSpotlight.stories.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { Main } from '../../abstracts';
+import { Main } from '../../../abstracts';
+import { ArticleSpotlight } from '../ArticleSpotlight';
 
-import { ArticleSpotlight } from './ArticleSpotlight';
+import thumbnail from './thumbnail.jpg';
 
 export default {
   component: ArticleSpotlight,
@@ -24,6 +25,5 @@ Primary.args = {
   description: `Quas explicabo a ad. Ut commodi est *iusto eaque et reiciendis* ullam fugiat. Ex inventore fuga.
 
   Neque porro quisquam est qui dolorem **ipsum quia dolor** sit amet, consectetur, adipisci velit.`,
-  thumbnailSrc:
-    'https://images.unsplash.com/photo-1615678815958-5910c6811c25?crop=entropy&cs=tinysrgb&fit=crop&fm=jpg&h=900&ixid=MnwxfDB8MXxyYW5kb218MHx8Y2F0fHx8fHx8MTY2MjkyMzI4NQ&ixlib=rb-1.2.1&q=80&w=1600',
+  thumbnailSrc: thumbnail,
 };