+++ packages/tokens/src/lib/colors/utils/getThemeColors.ts
@@ -4,7 +4,7 @@ export const getThemeColors = (colorTheme: ColorTheme, theme: Theme) => {
   return Object.keys(colorTheme[theme])
     .map((color) => {
       return Object.keys(colorTheme[theme][color]).map((shade) => {
-        return `--color-${color}-${shade}: ${colorTheme[theme][color][shade]};`;
+        return `--mrd-color-${color}-${shade}: ${colorTheme[theme][color][shade]};`;
       });
     })
     .flat();