+++ packages/tokens/src/lib/typography/index.ts
@@ -1,15 +1,12 @@
-import { CSSProperties } from 'styled-components';
-
 export type FontSizes = 0 | 1 | 2 | 3 | 4 | 5 | 'inherit';
 
 export type FontWeights = 'regular' | 'medium' | 'bold';
 
 export type FontFamily = 'sans' | 'serif' | 'code' | 'ui' | 'inherit';
 
-export const fontSizes: Record<
-  FontSizes,
-  NonNullable<CSSProperties['fontSize']>
-> = {
+export const baseFontSize = 16;
+
+export const fontSizes: Record<FontSizes, string> = {
   0: '0.875rem', // 14px
   1: '1rem', // 16px
   2: '1.5rem', // 24px
@@ -19,19 +16,13 @@ export const fontSizes: Record<
   inherit: 'inherit',
 };
 
-export const fontWeights: Record<
-  FontWeights,
-  NonNullable<CSSProperties['fontWeight']>
-> = {
+export const fontWeights: Record<FontWeights, number> = {
   regular: 400,
   medium: 500,
   bold: 700,
 };
 
-export const fontFamilies: Record<
-  FontFamily,
-  NonNullable<CSSProperties['fontFamily']>
-> = {
+export const fontFamilies: Record<FontFamily, string> = {
   sans: `'DM Sans',sans-serif`,
   serif: `'Eczar',serif`,
   code: `'JetBrains Mono',monospace`,